requirements.txt
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.1
hawkauthlib==2.0.0
idna==3.3
PyBrowserID==0.14.0
pycparser==2.21
PyFxA==0.7.7
PyJWT==2.3.0
requests==2.27.1
six==1.16.0
urllib3==1.26.8
WebOb==1.8.7
